fix: improves Gazebo reset behavoir
This commits improves the Gazebo reset behavoir. In the new form the
previous pose before the reset is saved and applied again after the
reset. This was done because of a bug in Gazebo which caused the reset
behavoir to differ between differen physics engines (see
https://github.com/osrf/gazebo/issues/3150). Further the `_set_init_pose`
method is now called after the reset.