feat(panda): improve init pose set behavoir
This commit improves the initial pose set behavoir. It fixes a bug in the application
of the init pose boundary regions. It further adds the `pose_sampling_type` parameter
to the task environment configuration file. This parameter can be used
to change the initial position sampling type. The supported types are:
`end_effector_pose` and `joint_positions`. The first one samples the
init pose from a set of EE poses where the latter samples a init pose
based on a set of joint positions.